+++ src/redirect.c
@@ -7,71 +7,27 @@ bool	redirect(t_token **token)
 		return (true);
 	if ((*token)->fd < 0)
 		return (false);
-	if ((*token)->type >= REDIRECT_IN && (*token)->type <= REDIRECT_APPEND)
+	if ((*token)->type == REDIRECT_IN || (*token)->type == REDIRECT_HERE_DOC)
 	{
-		if ((*token)->type == REDIRECT_IN || (*token)->type == REDIRECT_HERE_DOC)
-		{
-			if (((*token)->backup_fd = dup(STDIN_FILENO)) == -1)
-				ft_error("minishell", NULL, "dup failed", 1);
-			if (dup2((*token)->fd, STDIN_FILENO) == -1)
-				ft_error("minishell", NULL, "dup2 failed", 1);
-		}
-		else
-		{
-			if (((*token)->backup_fd = dup(STDOUT_FILENO)) == -1)
-				ft_error("minishell", NULL, "dup failed", 1);
-			if (dup2((*token)->fd, STDOUT_FILENO) == -1)
-				ft_error("minishell", NULL, "dup2 failed", 1);
-		}
-		if (close((*token)->fd) == -1)
-			ft_error("minishell", NULL, "cfase failed", 1);
-		*token = (*token)->next;
-		*token = (*token)->next;
+		(*token)->backup_fd = dup(STDIN_FILENO);
+		if ((*token)->backup_fd == -1)
+			ft_error("minishell", NULL, "dup failed", 1);
+		if (dup2((*token)->fd, STDIN_FILENO) == -1)
+			ft_error("minishell", NULL, "dup2 failed", 1);
 	}
+	else
+	{
+		(*token)->backup_fd = dup(STDOUT_FILENO);
+		if ((*token)->backup_fd == -1)
+			ft_error("minishell", NULL, "dup failed", 1);
+		if (dup2((*token)->fd, STDOUT_FILENO) == -1)
+			ft_error("minishell", NULL, "dup2 failed", 1);
+	}
+	if (close((*token)->fd) == -1)
+		ft_error("minishell", NULL, "cfase failed", 1);
+	*token = (*token)->next;
+	*token = (*token)->next;
 	return (true);
-}
-
-void	read_heredoc(char *eof, int tmp_fd)
-{
-	char *line;
-
-	while (1)
-	{
-		line = readline("> ");
-		if (!line || *line == '\n' || (ft_strlen(line) == ft_strlen(eof) &&
-			!ft_strncmp(line, eof, ft_strlen(line))))
-		{
-			if (!line)
-			{
-				ft_error("minishell", "warning",
-					"here-document delimited by end-of-file", 0);
-				break ;
-			}
-			free(line);
-			if (*line == '\n')
-				continue ;
-			break ;
-		}
-		ft_putendl_fd(line, tmp_fd);
-		free(line);
-	}
-}
-
-int	open_heredoc(char *eof)
-{
-	int tmp_fd;
-	int fd;
-
-	tmp_fd = open(HEREDOC, O_WRONLY | O_CREAT | O_TRUNC, S_IREAD | S_IWRITE);
-	if (tmp_fd == -1)
-		ft_error("minishell", "here_doc", "No such file or directory", 0);
-	read_heredoc(eof, tmp_fd);
-	close(tmp_fd);
-	fd = open(HEREDOC, O_RDONLY);
-	if (fd == -1)
-		ft_error("minishell", "here_doc", "No such file or directory", 0);
-	unlink(HEREDOC);
-	return (fd);
 }
 
 void	redirect_open(t_token *token)
@@ -81,22 +37,23 @@ void	redirect_open(t_token *token)
 		if (token->type == REDIRECT_IN)
 			token->fd = open(token->next->str, O_RDONLY);
 		else if (token->type == REDIRECT_OUT)
-			token->fd = open(token->next->str, O_WRONLY | O_CREAT | O_TRUNC, S_IREAD | S_IWRITE);
+			token->fd = open(token->next->str,
+					O_WRONLY | O_CREAT | O_TRUNC, S_IREAD | S_IWRITE);
 		else if (token->type == REDIRECT_APPEND)
-			token->fd = open(token->next->str, O_WRONLY | O_CREAT | O_APPEND, S_IREAD | S_IWRITE);
+			token->fd = open(token->next->str,
+					O_WRONLY | O_CREAT | O_APPEND, S_IREAD | S_IWRITE);
 		else if (token->type == REDIRECT_HERE_DOC)
 			token->fd = open_heredoc(token->next->str);
 		if (token->fd == -1)
-			ft_error("minishell", token->next->str, "No such file or directory", 0);
+			ft_error("minishell", token->next->str,
+				"No such file or directory", 0);
 		token = token->next;
 	}
 }
 
-
-// backup_fdがないと完成しない
 void	close_redirect(t_token *token)
 {
-	t_token *tmp;
+	t_token	*tmp;
 
 	tmp = token;
 	while (token->next)
